🚚 Move overview info to patterns
