+++ src/patterns/index.ts
@@ -1,3 +1,4 @@
 export * from "./Footer/Footer"
 export * from "./Navigation/Navigation"
+export * from "./OverviewInfo/OverviewInfo"
 export * from "./SayHi/SayHi"
